prueba-tecnica/telecom-backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
mysqlclient
passlib[bcrypt]
python-dotenv
